requirements.txt
selenium~=4.27.1
pytest~=8.3.4
allure-python-commons~=2.13.5